+++ solver/src/test/java/org/chocosolver/solver/constraints/nary/RegularTest.java
@@ -9,9 +9,8 @@
  */
 package org.chocosolver.solver.constraints.nary;
 
-import dk.brics.automaton.Automaton;
-import dk.brics.automaton.RegExp;
 import gnu.trove.set.hash.TIntHashSet;
+
 import org.chocosolver.solver.Model;
 import org.chocosolver.solver.Solution;
 import org.chocosolver.solver.Solver;
@@ -29,6 +28,9 @@ import java.lang.reflect.InvocationTargetException;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+
+import dk.brics.automaton.Automaton;
+import dk.brics.automaton.RegExp;
 
 import static org.chocosolver.solver.search.strategy.Search.inputOrderLBSearch;
 import static org.testng.Assert.assertEquals;
@@ -222,10 +224,10 @@ public class RegularTest {
         }
 
         assertEquals(1, solutions.size());
-        assertEquals(-9, solutions.get(0).getIntVal(CS[0]));
-        assertEquals(1, solutions.get(0).getIntVal(CS[1]));
-        assertEquals(1, solutions.get(0).getIntVal(CS[2]));
-        assertEquals(-5, solutions.get(0).getIntVal(CS[3]));
+        assertEquals(-9, (int) solutions.get(0).getIntVal(CS[0]));
+        assertEquals(1, (int) solutions.get(0).getIntVal(CS[1]));
+        assertEquals(1, (int) solutions.get(0).getIntVal(CS[2]));
+        assertEquals(-5, (int) solutions.get(0).getIntVal(CS[3]));
     }
 
     @Test(groups="1s", timeOut=60000)
